Rename linenumber file
