+++ spyder/widgets/panels/linenumber.py
