+++ spyder/widgets/sourcecode/codeeditor.py
@@ -57,7 +57,7 @@ from spyder.utils.sourcecode import ALL_LANGUAGES, CELL_LANGUAGES
 from spyder.widgets.editortools import PythonCFM
 from spyder.widgets.sourcecode.base import TextEditBaseWidget
 from spyder.widgets.sourcecode.kill_ring import QtKillRing
-from spyder.widgets.panels.line_number import LineNumberArea
+from spyder.widgets.panels.linenumber import LineNumberArea
 from spyder.widgets.panels.edgeline import EdgeLine
 from spyder.widgets.panels.manager import PanelsManager
